Extract cyclonedx conversion helpers to pkg/sbom/convert
