+++ pkg/sbom/convert/convert.go
@@ -5,7 +5,7 @@
 
 //go:build trivy || (windows && wmi)
 
-package sbom
+package convert
 
 import (
 	"time"
@@ -15,7 +15,6 @@ import (
 
 	"github.com/CycloneDX/cyclonedx-go"
 	"github.com/DataDog/agent-payload/v5/cyclonedx_v1_4"
-	"google.golang.org/protobuf/types/known/durationpb"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
@@ -852,10 +851,6 @@ func convertTimestamp(in string) *timestamppb.Timestamp {
 	}
 
 	return timestamppb.New(ts)
-}
-
-func convertDuration(in time.Duration) *durationpb.Duration {
-	return durationpb.New(in)
 }
 
 func convertTool(in *cyclonedx.Tool) *cyclonedx_v1_4.Tool { //nolint:staticcheck
